Revert "refactor: Use uint80 instead of uint64 for consistency"
This reverts commit e4a323cdb11ed312030d4c02da5539d9b6d941a8.

+++ test/foundry/Raffle.claimFees.t.sol
@@ -35,9 +35,9 @@ contract Raffle_ClaimFees_Test is TestHelpers {
         vm.startPrank(user1);
         looksRareRaffle.createRaffle({
             cutoffTime: uint40(block.timestamp + 86_400),
-            minimumEntries: uint80(107),
-            maximumEntries: uint80(200),
-            maximumEntriesPerParticipant: uint80(100),
+            minimumEntries: uint64(107),
+            maximumEntries: uint64(200),
+            maximumEntriesPerParticipant: uint64(100),
             prizesTotalValue: 1 ether,
             minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
